adds exp multi to entity configs, fixes currency tooltips, change commands to /exile

+++ src/main/java/com/robertx22/mine_and_slash/gui/overlays/bar_overlays/types/VanillaOverlay.java
@@ -188,7 +188,7 @@ public class VanillaOverlay extends DrawableHelper implements HudRenderCallback 
         if (max < 1) {
             return;
         }
-        int tenth = max / 10;
+        float tenth = max / 10F;
 
         int X_SPACING = 8;
